Implement Cast

+++ simple_model.py
@@ -22,7 +22,7 @@ class Model(nn.Module):
         # self.blocks[0].weight.data.fill_(1.2)
 
     def forward(self, x):
-        return self.blocks(x)
+        return self.blocks(x).int()
         # return x.permute(1, 0)
 
 model = Model()
